added wallet integration test and ready for release
