+++ packages/cryptography/package.json
@@ -25,6 +25,7 @@
         "./src/primitive/aes.js": "./src/primitive/aes.browser.js",
         "./src/primitive/sha256.js": "./src/primitive/sha256.browser.js",
         "./src/primitive/pbkdf2.js": "./src/primitive/pbkdf2.browser.js",
+        "./src/primitive/random.js": "./src/primitive/random.js",
         "./src/primitive/hmac.js": "./src/primitive/hmac.browser.js",
         "./src/primitive/bip32.js": "./src/primitive/bip32.browser.js"
     },
